move db_functions to models.functions
